Move tests to own files
